Add files via upload
Logically ordering game information from the video game series Advanced Wars. Using an OOP approach to programming there is a hierarchical structure to the data included and this is exemplified by the construction of the classes. Sub-classing the data allows for the logical differences in the instances to be localised, whilst shared attributes can avoid being repeated. 
Polymorphism in the form of method-overriding is shown briefly in the case of the "__repr__" method, with the same method name occupying difference namespaces based on local association within it's respective class.